fix: cast halfh to integer to avoid crop errors when adjusting aspect ratio

+++ mpsfm/extraction/pairwise/models/mast3r.py
@@ -99,7 +99,7 @@ def load_images(folder_or_list, size, square_ok=False, verbose=True):
         else:
             halfw, halfh = ((2 * cx) // 16) * 8, ((2 * cy) // 16) * 8
             if not (square_ok) and W == H:
-                halfh = 3 * halfw / 4
+                halfh = int(3 * halfw / 4)
             img = img.crop((cx - halfw, cy - halfh, cx + halfw, cy + halfh))
 
         W2, H2 = img.size
@@ -230,7 +230,7 @@ class Mast3rMatcher(BaseModel):
         cx, cy = W // 2, H // 2
         halfw, halfh = ((2 * cx) // 16) * 8, ((2 * cy) // 16) * 8
         if not (square_ok) and W == H:
-            halfh = 3 * halfw / 4
+            halfh = int(3 * halfw / 4)
         image = image[..., cy - halfh : cy + halfh, cx - halfw : cx + halfw]
         H2, W2 = image.shape[-2:]
         ImgNorm = tvf.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
